*** empty log message *** Original: d6ac926318ecf2b9e1f4f5cb35e357f28ea3e68c

+++ lalapps/src/stochastic/stochastic_new.c
@@ -1632,12 +1632,12 @@ void parseOptions(INT4 argc, CHAR *argv[])
 
       case 'k':
 	       /* high pass knee filter frequency  */
-	       highPassFreq= atoi(optarg);
+	       highPassFreq= atof(optarg);
 	       break;
                           
       case 'p':
 	       /* high pass filter attenuation  */
-	       highPassAt = atoi(optarg);
+	       highPassAt = atof(optarg);
 	       break;
 
       case 'P':
